+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v8.2.0] - 2020-07-31
+### Added
+- Add variable allowing different machine types
+
 ## [v8.1.0] - 2020-07-23
 ### Added
 - Add optional (turned on by `add_random_suffix` parameter) random suffix to all resources that will prevent more instances of this module to be provisioned